+++ ext/cuda/limiters.jl
@@ -1,5 +1,11 @@
-import ClimaCore.Limiters: QuasiMonotoneLimiter
+import ClimaCore.Limiters:
+    QuasiMonotoneLimiter,
+    compute_element_bounds!,
+    compute_neighbor_bounds_local!,
+    apply_limiter!
 import ClimaCore.Fields
+import ClimaCore: DataLayouts, Spaces, Topologies, Fields
+using CUDA
 
 function config_threadblock(Nv, Nh)
     nitems = Nv * Nh
