+++ ext/cuda/operators_finite_difference.jl
@@ -3,7 +3,9 @@ import Base.Broadcast: Broadcasted
 import ClimaComms
 using CUDA: @cuda
 import ClimaCore.Operators: AbstractStencilStyle, strip_space
+import ClimaCore.Operators: setidx!, getidx
 import ClimaCore.Operators: StencilBroadcasted
+import ClimaCore.Operators: LeftBoundaryWindow, RightBoundaryWindow, Interior
 
 struct CUDAColumnStencilStyle <: AbstractStencilStyle end
 AbstractStencilStyle(::ClimaComms.CUDADevice) = CUDAColumnStencilStyle
@@ -24,11 +26,11 @@ function Base.copyto!(
         Nq = 1
         Nh = 1
     end
-    (li, lw, rw, ri) = bounds = window_bounds(space, bc)
+    (li, lw, rw, ri) = bounds = Operators.window_bounds(space, bc)
     Nv = ri - li + 1
     max_threads = 256
     nitems = Nv * Nq * Nq * Nh # # of independent items
-    (nthreads, nblocks) = Topologies._configure_threadblock(max_threads, nitems)
+    (nthreads, nblocks) = _configure_threadblock(max_threads, nitems)
     @cuda always_inline = true threads = (nthreads,) blocks = (nblocks,) copyto_stencil_kernel!(
         strip_space(out, space),
         strip_space(bc, space),
@@ -55,7 +57,13 @@ function copyto_stencil_kernel!(out, bc, space, bds, Nq, Nh, Nv)
                 RightBoundaryWindow{Spaces.right_boundary_name(space)}() :
                 Interior()
             )
-        setidx!(space, out, idx, hidx, getidx(space, bc, window, idx, hidx))
+        setidx!(
+            space,
+            out,
+            idx,
+            hidx,
+            Operators.getidx(space, bc, window, idx, hidx),
+        )
     end
     return nothing
 end
