+++ ext/cuda/operators_thomas_algorithm.jl
@@ -1,11 +1,12 @@
 import ClimaComms
 import ClimaCore: Fields, Topologies, Spaces, Operators
-import ClimaCore.Operators: column_thomas_solve!
+import ClimaCore.Operators:
+    column_thomas_solve!, thomas_algorithm_kernel!, thomas_algorithm!
 import CUDA
 using CUDA: @cuda
 function column_thomas_solve!(::ClimaComms.CUDADevice, A, b)
     Ni, Nj, _, _, Nh = size(Fields.field_values(A))
-    nthreads, nblocks = Topologies._configure_threadblock(Ni * Nj * Nh)
+    nthreads, nblocks = _configure_threadblock(Ni * Nj * Nh)
     @cuda always_inline = true threads = nthreads blocks = nblocks thomas_algorithm_kernel!(
         A,
         b,
